+++ src/actions/remove.ts
@@ -1,7 +1,7 @@
 import { remove } from '../modules/fileTransferTask';
 import createFileAction from './createFileAction';
 
-export const removeRemote = createFileAction((source, config, { remoteFs }) =>
+export const removeRemote = createFileAction('remove', (source, config, { remoteFs }) =>
   remove(source, remoteFs, {
     ignore: config.ignore,
     skipDir: config.skipDir,