+++ src/commands/FileCommand.ts
@@ -31,7 +31,7 @@ export default class FileCommand extends Command {
     try {
       const config = getConfig(activityPath);
       await handler(activityPath, config);
-      logger.info(`${this.getName()} ${fileTarget.fsPath}`);
+      logger.info(`${this.getName()} ${activityPath}`);
     } catch (error) {
       output.onError(error);
     }